+++ .github/workflows/publish.yaml
@@ -32,3 +32,4 @@ jobs:
           push: true
           tags: ${{ steps.meta.outputs.tags }}
           labels: ${{ steps.meta.outputs.labels }}
+          build-args: RELEASE_VERSION=${{ steps.meta.outputs.version }}
