+++ constant/constant.go
@@ -1,7 +1,10 @@
 package constant
 
+// Version is the current versions of kwatch
+const Version = "dev"
+
 // WelcomeMsg is used to be sent to all providers when kwatch starts
-const WelcomeMsg = ":tada: kwatch just started!"
+const WelcomeMsg = ":tada: kwatch@%s just started!"
 
 // NumRequeues indicates number of retries when worker fails to handle item
 const NumRequeues = 5
