+++ go.mod
@@ -3,6 +3,7 @@ module github.com/abahmed/kwatch
 go 1.17
 
 require (
+	github.com/bwmarrin/discordgo v0.23.2
 	github.com/sirupsen/logrus v1.8.1
 	github.com/slack-go/slack v0.10.0
 	github.com/spf13/viper v1.9.0
@@ -12,7 +13,6 @@ require (
 )
 
 require (
-	github.com/bwmarrin/discordgo v0.23.2 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/fsnotify/fsnotify v1.5.1 // indirect
 	github.com/go-logr/logr v0.4.0 // indirect
